+++ Source/Nett.Exp/TomlValueExtensions.cs
@@ -1,0 +1,8 @@
+﻿namespace Nett.Exp
+{
+    public static class TomlValueExtensions
+    {
+        public static string GetUnit(this TomlValue value, string def = "")
+            => value.Unit ?? def;
+    }
+}