+++ Test/Nett.Tests.Internal/Parser/CreateTableTests.cs
@@ -73,7 +73,7 @@ x = 1
         {
             var lexer = new Lexer(input);
             var tokens = lexer.Lex();
-            var parser = new Nett.Parser.Parser(new ParseInput(tokens));
+            var parser = new Nett.Parser.Parser(new ParseInput(tokens), TomlSettings.DefaultInstance);
             return parser.Parse().SyntaxNodeOrDefault().CreateTable();
         }
     }